Update pco2_Graffiti_Kanekal_Braun.py
I think editing code like this is tricky. The long lines of fd's and left's and right's are necessary to get a particular drawing, so the clunkiness of this code is slightly warranted. So, I think, in the broad strokes, there are a couple things I'd suggest. More commenting so I know what's what. Assume I was coding this; if I was, would I have any idea what you were doing? Hopefully the answer to that question is YES! I mentioned in one part of your code that you could use a for loop. I think that's valid, but, again, it's difficult because of the particulars of this assignment. For instance, I don't think you had to do stuff so much repeatedly (thus warranting a for loop) but had to do particular strokes one at a time, which doesn't really work with for loops. 
All in all, I think you did a good job! I have few complaints. Just comment, comment, comment. It's a pain, I know. But, I had no idea what was doing what, and I think that's still really really important. 
I hope this helps.